database and distance input now works, adjusted some form stuff

+++ models/schema.sql
@@ -26,6 +26,6 @@ CREATE TABLE IF NOT EXISTS users (
 */
 CREATE TABLE IF NOT EXISTS walks (
     id INTEGER NOT NULL,
-    distance INTEGER NOT NULL,
+    distance REAL NOT NULL,
     walkdate TEXT NOT NULL
 );
